NeuralChat/backend/requirements.txt
azure-functions==1.21.3
fastapi==0.115.6
uvicorn==0.32.1
httpx==0.28.1
pydantic==2.10.3
